Added option to export packs as folders

+++ common/src/main/java/dev/latvian/mods/kubejs/command/KubeJSCommands.java
@@ -59,12 +59,15 @@ import net.minecraft.world.InteractionHand;
 import net.minecraft.world.item.ItemStack;
 import org.apache.commons.io.FileUtils;
 
+import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Modifier;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -137,8 +140,11 @@ public class KubeJSCommands {
 			.then(Commands.literal("export")
 				.requires(source -> source.getServer().isSingleplayer() || source.hasPermission(2))
 				.executes(context -> export(context.getSource()))
-				.then(Commands.literal("packs")
-					.executes(context -> exportPacks(context.getSource()))
+				.then(Commands.literal("pack_zips")
+					.executes(context -> exportPacks(context.getSource(), true))
+				)
+				.then(Commands.literal("pack_folders")
+					.executes(context -> exportPacks(context.getSource(), false))
 				)
 			)
 			/*
@@ -649,7 +655,7 @@ public class KubeJSCommands {
 		return 1;
 	}
 
-	private static int exportPacks(CommandSourceStack source) {
+	private static int exportPacks(CommandSourceStack source, boolean exportZip) {
 		var packs = new ArrayList<ExportablePackResources>();
 
 		for (var pack : source.getServer().getResourceManager().listPacks().toList()) {
@@ -662,15 +668,29 @@ public class KubeJSCommands {
 		int success = 0;
 
 		for (var pack : packs) {
-			var path = KubeJSPaths.EXPORTED_PACKS.resolve(pack.packId() + ".zip");
 			try {
-				Files.deleteIfExists(path);
+				if (exportZip) {
+					var path = KubeJSPaths.EXPORTED_PACKS.resolve(pack.packId() + ".zip");
+					Files.deleteIfExists(path);
 
-				try (var fs = FileSystems.newFileSystem(path, Map.of("create", true))) {
-					pack.export(fs);
+					try (var fs = FileSystems.newFileSystem(path, Map.of("create", true))) {
+						pack.export(fs.getPath("."));
+					}
+				} else {
+					var path = KubeJSPaths.EXPORTED_PACKS.resolve(pack.packId());
+
+					if (Files.exists(path)) {
+						Files.walk(path)
+							.sorted(Comparator.reverseOrder())
+							.map(Path::toFile)
+							.forEach(File::delete);
+					}
+
+					Files.createDirectories(path);
+					pack.export(path);
 				}
 
-				source.sendSuccess(() -> Component.literal("Successfully exported ").withStyle(ChatFormatting.GREEN).append(Component.literal(pack.packId()).withStyle(ChatFormatting.BLUE)), true);
+				source.sendSuccess(() -> Component.literal("Successfully exported ").withStyle(ChatFormatting.GREEN).append(Component.literal(pack.packId()).withStyle(ChatFormatting.BLUE)), false);
 				success++;
 			} catch (IOException e) {
 				e.printStackTrace();
@@ -678,6 +698,14 @@ public class KubeJSCommands {
 					style.withColor(ChatFormatting.RED)
 						.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal(e.getMessage())))));
 			}
+		}
+
+		int success1 = success;
+
+		if (source.getServer().isSingleplayer() && !source.getServer().isPublished()) {
+			source.sendSuccess(() -> Component.literal("Exported " + success1 + " packs").kjs$clickOpenFile(KubeJSPaths.EXPORTED_PACKS.toAbsolutePath().toString()), false);
+		} else {
+			source.sendSuccess(() -> Component.literal("Exported " + success1 + " packs"), false);
 		}
 
 		return success;
